+++ html-router/src/lib.rs
@@ -36,6 +36,7 @@ where
         .add_protected_routes(routes::content::router())
         .add_protected_routes(routes::knowledge::router())
         .add_protected_routes(routes::ingestion::router())
+        .add_protected_routes(routes::scratchpad::router())
         .with_compression()
         .build()
 }